add observer pattern

+++ src/main/java/org/example/model/dto/CommentDTORecord.java
@@ -1,3 +1,3 @@
 package org.example.model.dto;
 
-public record CommentDTORecord(Long id, Long postId, Long authorId, String text) {}
+public record CommentDTORecord(Integer id, Integer postId, Long authorId, String text) {}
